+++ internal/authentification/auth.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"final_project/internal/common"
 	"net/http"
-	"os"
 
 	"github.com/golang-jwt/jwt/v5"
 )
@@ -19,7 +18,7 @@ func Auth(next http.HandlerFunc) http.HandlerFunc {
 			return
 		}
 		key := []byte("key")
-		hashPassword := sha256.Sum256([]byte(os.Getenv("TODO_PASSWORD")))
+		hashPassword := sha256.Sum256([]byte(common.Password))
 		claims := jwt.MapClaims{"password": hashPassword}
 		jwtUns := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
 		jwtSign, err := jwtUns.SignedString(key)
