+++ internal/common/constants.go
@@ -1,0 +1,10 @@
+package common
+
+import "os"
+
+const TimeFormat = "20060102"
+
+const Limit = "10"
+
+var Password = os.Getenv("TODO_PASSWORD")
+
