--- internal/common/variables.go
@@ -1,3 +1,0 @@
-package common
-
-var TimeFormat = "20060102"
