--- internal/handlers/add_task/handler.go
@@ -1,11 +1,0 @@
-package addtask
-
-import "final_project/internal/repository"
-
-type Handler struct {
-	repo *repository.Repository
-}
-
-func New(repo *repository.Repository) *Handler {
-	return &Handler{repo: repo}
-}
