--- internal/handlers/delete_task/handler.go
@@ -1,13 +1,0 @@
-package deletetask
-
-import "final_project/internal/repository"
-
-type Handler struct {
-	rep *repository.Repository
-}
-
-func New(rep *repository.Repository) *Handler {
-	return &Handler{
-		rep: rep,
-	}
-}
