--- internal/handlers/done_task/handler.go
@@ -1,11 +1,0 @@
-package donetask
-
-import "final_project/internal/repository"
-
-type Handler struct {
-	rep *repository.Repository
-}
-
-func New(rep *repository.Repository) *Handler {
-	return &Handler{rep: rep}
-}
