+++ internal/handlers/handler.go
@@ -1,0 +1,11 @@
+package handlers
+
+import "final_project/internal/repository"
+
+type Handler struct {
+	rep *repository.Repository
+}
+
+func New(rep *repository.Repository) *Handler {
+	return &Handler{rep: rep}
+}
