+++ internal/repository/delete_task.go
@@ -1,12 +1,18 @@
 package repository
 
-import "database/sql"
+import (
+	"database/sql"
+	"errors"
+)
 
 func (rep *Repository) DeleteTask(id string) error {
 	query := "DELETE FROM scheduler WHERE id=:id"
-	_, err := rep.db.Exec(query, sql.Named("id", id))
+	res, err := rep.db.Exec(query, sql.Named("id", id))
 	if err != nil {
 		return err
+	}
+	if num, _ := res.RowsAffected(); num == 0 {
+		return errors.New("задача не найдена")
 	}
 	return nil
 }
