+++ tests/settings.go
@@ -1,9 +1,7 @@
 package tests
 
-import "os"
-
 var Port = 7540
-var DBFile = os.Getenv("TODO_DBFILE")
+var DBFile = "./scheduler.db"
 var FullNextDate = false
 var Search = true
 var Token = `eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJwYXNzd29yZCI6WzE4LDIzMiw5NCwzNiwxNzgsMjAwLDE3MiwxMiwxMTQsMTg1LDc3LDE0LDEyLDU0LDk2LDE3MCwyMTEsMTQwLDE3MiwxNzgsMjA1LDIxNSwxNjEsMTQ0LDYxLDEzNSw1MCwyNDUsMTYwLDI0NCwxMzQsNjBdfQ.h9GJnmvY6fWslCZTppQ3LMYuIxHsQj9fFEdv3p7gZUc`
